use Theme hook layout page more clean move files again*
