+++ src/components/UIProvider.tsx
@@ -1,4 +1,4 @@
 import { NextUIProvider } from "@nextui-org/react";
-export function NextuiProvider({ children }: { children: React.ReactNode }) {
+export function UIProvider({ children }: { children: React.ReactNode }) {
   return <NextUIProvider>{children}</NextUIProvider>;
 }
