+++ src/lib/stores/useThemeStore.tsx
@@ -1,0 +1,11 @@
+import { create } from "zustand";
+
+interface ThemeState {
+  theme: "light" | "dark";
+  setTheme: (theme: "light" | "dark") => void;
+}
+
+export const useThemeStore = create<ThemeState>((set) => ({
+  theme: "light",
+  setTheme: (theme) => set({ theme }),
+}));
