--- src/lib/utils/testing.tsx
@@ -1,9 +1,0 @@
-"use client";
-const systemTheme = () => {
-  const prefersDarkMode = window.matchMedia(
-    "(prefers-color-scheme: dark)"
-  ).matches;
-  return prefersDarkMode ? "dark" : "light";
-};
-
-export default systemTheme;
